journal/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Redis for pub/sub
redis==4.6.0

# Socket.IO server
python-socketio==5.8.0
eventlet==0.33.3

# Authentication and security
cryptography==41.0.4
bcrypt==4.0.1

# Data validation
marshmallow==3.20.1
jsonschema==4.19.1

# Additional dependencies
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7

# Financial data
yfinance==0.2.18
pandas==2.0.3
numpy==1.24.3

# Production server
gunicorn==21.2.0
gevent==23.7.0
gevent-websocket==0.10.1

# Web scraping (for prop firm scraper)
beautifulsoup4==4.12.2
selenium==4.11.2

# Email service
mailchimp-marketing==3.0.80

# Additional utilities
urllib3==2.0.4
stripe==5.5.0
